requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
cloudinary==1.34.0
Django==4.2.5
django-ckeditor==6.7.0
django-ckeditor-5==0.2.9
django-cloudinary-storage==0.3.0
django-js-asset==2.1.0
idna==3.4
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0
